backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.5",
    "pinecone>=7.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "uvicorn>=0.40.0",
    "twilio>=8.0.0",
    "python-multipart>=0.0.9",
]
